requirements.txt

firebase-admin==6.5.0
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud==0.34.0
google-cloud-aiplatform==1.71.1
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-logging==3.11.3
google-cloud-pubsub==2.28.0
grpcio==1.64.1
grpcio-status==1.62.2
protobuf==4.25.3
rsa==4.9
uritemplate==4.1.1
urllib3==2.2.2
